Components js and scss
