+++ src/components/banner.js
@@ -1,0 +1,20 @@
+function banner() {
+  return (
+    <section className="banner">
+      <div className="banner__iam">
+        <div className="banner__text">
+          Soy&nbsp;
+          <span className="banner__text-write">una creativa.</span>
+        </div>
+        <p className="banner__text">desarrollara Frontend</p>
+      </div>
+      <section className="banner__image">
+        <section className="banner__image-imageprofile">
+          <div className="banner__image-imageMe"></div>
+        </section>
+      </section>
+    </section>
+  );
+}
+
+export default banner;
