+++ src/components/menutablet.js
@@ -1,0 +1,23 @@
+function menutablet() {
+  return (
+    <nav className="menu__tablet">
+      <a className="menu__sections" href="#contactme">
+        contact
+      </a>
+      <a className="menu__sections" href="#profile">
+        portafolio
+      </a>
+      <a className="menu__sections" href="#experience">
+        experience
+      </a>
+      <a className="menu__sections" href="#skills">
+        skills
+      </a>
+      <a className="menu__sections" href="#aboutme">
+        about_me
+      </a>
+    </nav>
+  );
+}
+
+export default menutablet;
