+++ src/styles/core/variables.scss
@@ -1,0 +1,11 @@
+// sizes
+$mobile-width: 320px;
+$tablet-width: 768px;
+$desktop-width: 1200px;
+
+$color-white: white;
+$color-black: black;
+$color-brown: rgb(230, 161, 71);
+$color-pink: #ca3ea6;
+$color-blue: #3972b8;
+$color-grey: rgb(156, 150, 150);
