--- src/styles/reset.scss
@@ -1,9 +1,0 @@
-* {
-  box-sizing: border-box;
-  margin: 0;
-  padding: 0;
-  border-style: none;
-  text-decoration: none;
-  list-style: none;
-  color: white;
-}
